user who did not create posts cannot update and delete

+++ backend/controllers/posts.js
@@ -6,21 +6,21 @@ exports.getPosts = (req, res, next) => {
     const postQuery = Post.find();
     let fetchedPosts;
     if (pageSize && currentPage) {
-      postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
+        postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
     }
     postQuery
-      .then(documents => {
-        fetchedPosts = documents;
-        return Post.countDocuments();
-      })
-      .then(count => {
-        res.status(200).json({
-          message: "Posts fetched successfully!",
-          posts: fetchedPosts,
-          maxPosts: count
+        .then(documents => {
+            fetchedPosts = documents;
+            return Post.countDocuments();
+        })
+        .then(count => {
+            res.status(200).json({
+                message: "Posts fetched successfully!",
+                posts: fetchedPosts,
+                maxPosts: count
+            });
         });
-      });
-  };
+};
 
 exports.getPost = async (req, res, next) => {
     const postId = req.params.postId;
@@ -38,7 +38,8 @@ exports.postPosts = (req, res, next) => {
     const newPost = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     })
     newPost.save()
         .then(createdPost => {
@@ -56,8 +57,11 @@ exports.postPosts = (req, res, next) => {
 
 exports.deletePost = async (req, res, next) => {
     const postId = req.params.postId;
-    Post.deleteOne({ _id: postId })
+    Post.deleteOne({ creator: req.userData.userId, _id: postId })
         .then(result => {
+            if (!result) {
+                return res.status(404).json({ message: "Cannot delete posts you didn't create" })
+            }
             res.status(200).json({
                 message: 'post deleted'
             })
@@ -79,13 +83,16 @@ exports.editPost = (req, res, next) => {
         content: req.body.content,
         imagePath: imagePath
     }
-    Post.findById(postId)
+    Post.findOne({ creator: req.userData.userId, _id: postId })
         .then(post => {
+            if (!post) {
+                return res.status(404).json({ message: "Cannot edit post you didn't create" })
+            }
             post.title = newPost.title;
             post.content = newPost.content;
             post.imagePath = newPost.imagePath
             console.log(post)
-            return post.save()
+            post.save()
                 .then(result => {
                     res.status(200).json({ message: 'post updated' })
                 })
